로그아웃, refresh,블랙리스트 토큰 ,  팀별게시판생성기능구현 ,accessToken redis처리

+++ baseball-community-BE/src/main/java/com/baseball/baseballcommunitybe/post/entity/Post.java
@@ -31,7 +31,7 @@ public class Post {
 
     //  KBO 팀 id
     @Column(name = "team_id", nullable = false)
-    private Byte teamId;
+    private Long teamId;
 
     @Column(nullable = false, length = 200)
     private String title;
